PAI (Lab)/Task # 6/animal_herd_detection/requirement.txt
opencv-python
numpy
folium
geocoder
requests
gpsd-py3
